test_pfe/02-orchestration-agents-layer/cicd-agent/requirements.txt
python-dotenv>=1.0.1
groq>=0.12.0
pydantic>=2.10.0,<3
pyyaml>=6.0.2
requests>=2.32.0
jsonschema>=4.23.0
python-gitlab>=4.8.0
PyGithub>=2.5.0
